Improve docstrings and README.
